test(linter/plugins): conformance tests set `NODE_DISABLE_COLORS` in NPM script (#18493)
Pure refactor. Instead of setting `process.env.NODE_DISABLE_COLORS` in the JS code of conformance tester, set it in the NPM `conformance` script. This allows combining `index.ts` and `main.ts` - simplification.
